BVH: Refactor Node to fit in 32 bytes Aligning to CPU cache lines helps with performance

+++ BVH/bvh.hh
@@ -9,29 +9,29 @@
 
 namespace BVH {
 struct LeafInfo {
-  size_t index;
-  size_t num_tris;
+  uint32_t index;
+  uint32_t num_tris;
 };
 
 class BVH {
   // Reference:
   // https://jacco.ompf2.com/2022/04/13/how-to-build-a-bvh-part-1-basics/
 private:
-  size_t num_used_nodes_ = 0;
+  uint32_t num_used_nodes_ = 0;
   // TODO: Idea exponentially expanding array with max grow size
   std::vector<Node> nodes_;
 
-  void build_tree(size_t parent_node_index);
-  size_t get_new_node_index();
-  void update_node_bounds(size_t node_index);
+  void build_tree(uint32_t parent_node_index);
+  uint32_t get_new_node_index();
+  void update_node_bounds(uint32_t node_index);
 
 public:
   std::vector<Triangle> triangles;
   const std::vector<Node> &nodes = nodes_;
 
   void update_tree();
-  size_t count_leaf_triangles(size_t node_index = 0) const;
-  size_t count_leaf_nodes(size_t node_index = 0) const;
-  LeafInfo find_biggest_leaf(size_t node_index = 0) const;
+  uint32_t count_leaf_triangles(uint32_t node_index = 0) const;
+  uint32_t count_leaf_nodes(uint32_t node_index = 0) const;
+  LeafInfo find_biggest_leaf(uint32_t node_index = 0) const;
 };
 } // namespace BVH
